+++ tests/litellm/proxy/db/test_spend_update_queue.py
@@ -7,7 +7,7 @@ import pytest
 from fastapi.testclient import TestClient
 
 from litellm.proxy._types import Litellm_EntityType, SpendUpdateQueueItem
-from litellm.proxy.db.spend_update_queue import SpendUpdateQueue
+from litellm.proxy.db.db_transaction_queue.spend_update_queue import SpendUpdateQueue
 
 sys.path.insert(
     0, os.path.abspath("../../..")
